Add lib.english
